requirements.txt
django-dbbackup
djangorestframework
persiantools
python-dateutil
pandas
requests
django
reportlab
arabic_reshaper
python-bidi
GuardPyCaptcha
decouple
python-dotenv
dj-database-url
gunicorn
django-cors-headers
opencv-python
dj-database-url
python-dotenv